+++ CHANGELOG.md
@@ -34,8 +34,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * Add REST API endpoint for the questionnaire report generation
 * Add REST API endpoint for the patient's caregiver(s) relationship(s) (!222)
 * Add unique constraint to Relationship model (!655)
-* Generate a random uuid for the caregiver username(!267)
-* Add a "Manage Caregiver Access to Patient Data" page(!271)
+* Generate a random uuid for the caregiver username (!267)
+* Add a "Manage Caregiver Access to Patient Data" page (!271)
 
 ### Documentation